split common
